+++ lib/jsonut.dart
@@ -37,14 +37,16 @@ import 'dart:convert' as dart;
 /// - [JsonAny.from]
 /// - [JsonAny.tryFrom]
 extension type const JsonValue._(Object? _value) {
-  static T _parse<T extends JsonValue?>(String json) {
+  static T _parseAndCastAs<T extends JsonValue?>(String json) {
     final value = dart.json.decode(json);
-    return value is T
-        ? value
-        : throw FormatException(
-            'Invalid JSON value: expected $T but got ${value.runtimeType}',
-            json,
-          );
+    if (value is T) {
+      return value;
+    }
+    throw ArgumentError.value(
+      value,
+      'json',
+      'Decoded value is ${value.runtimeType}, expected $T',
+    );
   }
 
   /// Converts this object into a [JsonAny].
@@ -57,7 +59,7 @@ extension type const JsonValue._(Object? _value) {
 /// to a specific type, or to check its type. For each possible JSON type, there
 /// is a cooresponding set of methods to:
 ///
-/// - Convert the value to the specific type (e.g. [boolean], [booleanOr]).
+/// - Convert the value to the specific type (e.g. [boolean], [booleanOrNull]).
 /// - Check if the value is of the specific type (e.g. [isBool]).
 ///
 /// ## Example
@@ -88,7 +90,7 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Parses and returns the given [input] as a JSON value.
   ///
   /// If parsing fails, a [FormatException] is thrown.
-  factory JsonAny.parse(String input) => JsonValue._parse(input);
+  factory JsonAny.parse(String input) => JsonValue._parseAndCastAs(input);
 
   /// Tries to convert the given [value] to a [JsonValue].
   ///
@@ -109,13 +111,41 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Whether the value is `null`.
   bool get isNull => _value == null;
 
+  /// Returns the value cast to the given type [T].
+  ///
+  /// This method is useful when you know the type of the value, and want to
+  /// avoid the overhead of checking the type and casting it manually. For
+  /// example:
+  ///
+  /// ```dart
+  /// final class Employee {
+  ///   const Employee({required this.name, required this.age});
+  ///
+  ///   factory Employee.fromJson(JsonObject object) {
+  ///     return Employee(
+  ///       name: object['name'].as(),
+  ///       age: object['age'].as(),
+  ///     );
+  ///   }
+  ///
+  ///   final String name;
+  ///   final int age;
+  /// }
+  /// ```
+  T as<T extends JsonValue>() => _value as T;
+
+  /// Returns the value cast to the given type [T], or `null` otherwise.
+  T? asOrNull<T extends JsonValue>() {
+    return _value is T ? _value : null;
+  }
+
   /// Returns the value as a boolean.
   ///
   /// If the value is not a boolean, an error is thrown.
-  JsonBool boolean() => _value as JsonBool;
+  JsonBool boolean() => as();
 
   /// Returns the value as a boolean, or `null` if it is not a boolean.
-  JsonBool? booleanOr() => _value is JsonBool ? _value : null;
+  JsonBool? booleanOrNull() => asOrNull();
 
   /// Whether the value is a boolean.
   bool get isBool => _value is JsonBool;
@@ -123,10 +153,10 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Returns the value as a number.
   ///
   /// If the value is not a number, an error is thrown.
-  JsonNumber number() => _value as JsonNumber;
+  JsonNumber number() => as();
 
   /// Returns the value as a number, or `null` if it is not a number.
-  JsonNumber? numberOr() => _value is JsonNumber ? _value : null;
+  JsonNumber? numberOrNull() => asOrNull();
 
   /// Whether the value is a number.
   bool get isNumber => _value is JsonNumber;
@@ -134,10 +164,10 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Returns the value as a string.
   ///
   /// If the value is not a string, an error is thrown.
-  JsonString string() => _value as JsonString;
+  JsonString string() => as();
 
   /// Returns the value as a string, or `null` if it is not a string.
-  JsonString? stringOr() => _value is JsonString ? _value : null;
+  JsonString? stringOrNull() => asOrNull();
 
   /// Whether the value is a string.
   bool get isString => _value is JsonString;
@@ -145,10 +175,10 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Returns the value as an array.
   ///
   /// If the value is not an array, an error is thrown.
-  JsonArray array() => _value as JsonArray;
+  JsonArray array() => as();
 
   /// Returns the value as an array, or `null` if it is not an array.
-  JsonArray? arrayOr() => _value is JsonArray ? _value : null;
+  JsonArray? arrayOrNull() => asOrNull();
 
   /// Whether the value is an array.
   bool get isArray => _value is JsonArray;
@@ -156,10 +186,10 @@ extension type const JsonAny._(Object? _value) implements JsonValue {
   /// Returns the value as an object.
   ///
   /// If the value is not an object, an error is thrown.
-  JsonObject object() => _value as JsonObject;
+  JsonObject object() => as();
 
   /// Returns the value as an object, or `null` if it is not an object.
-  JsonObject? objectOr() => _value is JsonObject ? _value : null;
+  JsonObject? objectOrNull() => asOrNull();
 
   /// Whether the value is an object.
   bool get isObject => _value is JsonObject;
@@ -173,29 +203,27 @@ extension type const JsonBool(bool _value) implements JsonValue, bool {
   ///
   /// If parsing fails, or the result is not a boolean, a [FormatException] is
   /// thrown.
-  factory JsonBool.parse(String input) => JsonValue._parse(input);
+  factory JsonBool.parse(String input) => JsonValue._parseAndCastAs(input);
 }
 
 /// A zero-cost wrapper around a JSON number.
 ///
 /// This type exists to provide a subtype of [JsonValue] that is a number.
 extension type const JsonNumber(num _value) implements JsonValue, num {
-  /// Parses and returns the given [input] as a number.
+  /// Parses and casts the given [input] as a number.
   ///
-  /// If parsing fails, or the result is not a number, a [FormatException] is
-  /// thrown.
-  factory JsonNumber.parse(String input) => JsonValue._parse(input);
+  /// If parsing fails a [FormatException] is thrown.
+  factory JsonNumber.parse(String input) => JsonValue._parseAndCastAs(input);
 }
 
 /// A zero-cost wrapper around a JSON string.
 ///
 /// This type exists to provide a subtype of [JsonValue] that is a string.
 extension type const JsonString(String _value) implements JsonValue, String {
-  /// Parses and returns the given [input] as a string.
+  /// Parses and casts the given [input] as a string.
   ///
-  /// If parsing fails, or the result is not a string, a [FormatException] is
-  /// thrown.
-  factory JsonString.parse(String input) => JsonValue._parse(input);
+  /// If parsing fails a [FormatException] is thrown.
+  factory JsonString.parse(String input) => JsonValue._parseAndCastAs(input);
 }
 
 /// A zero-cost wrapper around a JSON array.
@@ -216,7 +244,7 @@ extension type const JsonArray._(List<JsonAny> _value)
   ///
   /// If parsing fails, or the result is not a array, a [FormatException] is
   /// thrown.
-  factory JsonArray.parse(String input) => JsonValue._parse(input);
+  factory JsonArray.parse(String input) => JsonValue._parseAndCastAs(input);
 }
 
 /// A zero-cost wrapper around a JSON object.
@@ -233,65 +261,31 @@ extension type const JsonArray._(List<JsonAny> _value)
 /// void main() {
 ///   const json = '{"name": "John Doe", "age": 42}';
 ///   final object = JsonObject.parse(json);
-///   print(object.string('name')); // John Doe
-///   print(object.number('age')); // 42
-///   print(object.stringOr('email')); // null
+///   print(object['name'].string()); // John Doe
+///   print(object['age].number()); // 42
+///   print(object['email'].stringOrNull()); // null
 /// }
 /// ```
-extension type const JsonObject._(Map<String, JsonAny?> _value)
-    implements JsonValue, Map<String, JsonAny?> {
+extension type const JsonObject._(Map<String, JsonAny> fields)
+    implements JsonValue, Map<String, JsonAny> {
   /// Returns and treats the given [value] as a JSON object.
   ///
   /// This is a zero-cost operation, and is provided as a convenience so that
   /// you can treat a map of [String] to [JsonValue] as a JSON object without
   /// needing to create a new instance.
   factory JsonObject(Map<String, JsonValue> value) {
-    return JsonObject._(value as Map<String, JsonAny?>);
+    return JsonObject._(value as Map<String, JsonAny>);
   }
 
-  /// Parses and returns the given [input] as an object.
+  /// Parses and casts the given [input] as an object.
   ///
-  /// If parsing fails, or the result is not an object, a [FormatException] is
-  /// thrown.
-  factory JsonObject.parse(String input) => JsonValue._parse(input);
+  /// If parsing fails a [FormatException] is thrown.
+  factory JsonObject.parse(String input) => JsonValue._parseAndCastAs(input);
 
-  /// Returns the value of the field as a boolean.
+  /// Returns a zero-cost wrapper for the field with the given [name].
   ///
-  /// If the field does not exist, or is not a boolean, an error is thrown.
-  JsonBool boolean(String key) => _value[key]!.boolean();
-
-  /// Returns the value of the field as a boolean, or `null` otherwise.
-  JsonBool? booleanOr(String key) => _value[key]?.booleanOr();
-
-  /// Returns the value of the field as a number.
-  ///
-  /// If the field does not exist, or is not a number, an error is thrown.
-  JsonNumber number(String key) => _value[key]!.number();
-
-  /// Returns the value of the field as a number, or `null` otherwise.
-  JsonNumber? numberOr(String key) => _value[key]?.numberOr();
-
-  /// Returns the value of the field as a string.
-  ///
-  /// If the field does not exist, or is not a string, an error is thrown.
-  JsonString string(String key) => _value[key]!.string();
-
-  /// Returns the value of the field as a string, or `null` otherwise.
-  JsonString? stringOr(String key) => _value[key]?.stringOr();
-
-  /// Returns the value of the field as an array.
-  ///
-  /// If the field does not exist, or is not an array, an error is thrown.
-  JsonArray array(String key) => _value[key]!.array();
-
-  /// Returns the value of the field as an array, or `null` otherwise.
-  JsonArray? arrayOr(String key) => _value[key]?.arrayOr();
-
-  /// Returns the value of the field as an object.
-  ///
-  /// If the field does not exist, or is not an object, an error is thrown.
-  JsonObject object(String key) => _value[key]!.object();
-
-  /// Returns the value of the field as an object, or `null` otherwise.
-  JsonObject? objectOr(String key) => _value[key]?.objectOr();
+  /// This method shadows the `[]` operator to provide a type-safe way to access
+  /// fields in the JSON object without needing to cast the result or check its
+  /// type or nullability.
+  JsonAny operator [](String name) => JsonAny._(fields[name]);
 }